Moved  `are_equal` to `utils/calculate.py`
