+++ simpa/utils/calculate.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: MIT
 
 
+from typing import Union
 import numpy as np
 import torch
 from scipy.interpolate import interp1d
@@ -224,3 +225,22 @@ def positive_gauss(mean, std) -> float:
         return positive_gauss(mean, std)
     else:
         return random_value
+
+
+def are_equal(obj1: Union[list, tuple, np.ndarray, object], obj2: Union[list, tuple, np.ndarray, object]) -> bool:
+    """Compare if two objects are equal. For lists, tuples and arrays, all entries need to be equal to return True.
+
+    :param obj1: The first object to compare. Can be of any type, but typically a list, numpy array, or scalar.
+    :type obj1: Union[list, tuple, np.ndarray, object]
+    :param obj2: The second object to compare. Can be of any type, but typically a list, numpy array, or scalar.
+    :type obj2: Union[list, tuple, np.ndarray, object]
+    :return: True if the objects are equal, False otherwise. For lists and numpy arrays, returns True only if all
+        corresponding elements are equal.
+    :rtype: bool
+    """
+    # Check if one object is numpy array or list
+    if isinstance(obj1, (list, np.ndarray, tuple)) or isinstance(obj2, (list, np.ndarray, tuple)):
+        return np.array_equal(obj1, obj2)
+    # For other types, use standard equality check which also works for lists
+    else:
+        return obj1 == obj2
